% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the CHIRPS&CHIRTS Downloader Shiny App}
\usage{
run_app()
}
\value{
Starts the Shiny app interface.
}
\description{
This function runs the integrated CHIRPS Shiny application.
}
\examples{
if (interactive()) {
  run_app()
}
}
